+++ wily/__init__.py
@@ -5,7 +5,7 @@ import colorlog
 import datetime
 
 
-__version__ = "0.9.0.dev"
+__version__ = "1.0.0"
 
 _handler = colorlog.StreamHandler()
 _handler.setFormatter(colorlog.ColoredFormatter("%(log_color)s%(message)s"))