+++ wily/config.py
@@ -2,7 +2,6 @@
 Configuration of wily
 
 TODO : Handle operator settings. Maybe a section for each operator and then pass kwargs to operators?
-TODO : Allow configuration of cache path (incase it needs to go in another folder)
 TODO : Better utilise default values and factory in @dataclass to replace DEFAULT_CONFIG
  and replace the logic in load() to set default values.
 """