+++ HISTORY.md
@@ -1,5 +1,19 @@
 # Release History
 
+## 1.0.0 (9th November 2018)
+
+* Build now compares existing git history with the cached history and only builds the missing revisions instead of building the entire index
+* Will check if .wily/ is not in .gitignore before running build
+* Improved documentation..
+
+## 0.9.0 (9th November 2018)
+
+
+* The build command now requires the target path, and supports multiple paths. Is no longer -t option, but an argument. This is to prevent the user from accidentally trying to scan venv's
+* Operators all have a default metric (lines-of-code, maintainability-index)
+* Report command by default will now display the default metrics in an index
+* Report command now accepts multiple metrics and adds them to the table
+
 ## 0.8.0 (7th November 2018)
 
 * Add support for relative paths in cache